fix: rename CardInfo component and fix warning about html struct
