+++ src/components/CardInfo.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom/dist';
 import Card from 'react-bootstrap/Card';
 import Badge from 'react-bootstrap/Badge';
-import styles from '../css/components/Card.module.css';
+import styles from '../css/components/CardInfo.module.css';
 
 const CardInfo = () => {
   const movePage = useNavigate();
@@ -16,15 +16,13 @@ const CardInfo = () => {
       <Card onClick={toDetailPage} className={styles.card}>
         <Card.Body>
           <Card.Title className={styles.card_title}>밥먹자!</Card.Title>
-          <Card.Text>
-            <h5>
-              <Badge bg="primary">
-                식사
-              </Badge>{' '}
-              <Badge bg="primary">
-                한식
-              </Badge>{' '}
-            </h5>
+          <Card.Text className={styles.badges}>
+            <Badge bg="primary">
+              식사
+            </Badge>{' '}
+            <Badge bg="primary">
+              한식
+            </Badge>{' '}
           </Card.Text>
           <Card.Text>
             2023/04/17 19:30
