+++ src/css/components/CardInfo.module.css
@@ -21,3 +21,7 @@
     margin-top:40px;
     text-align: end;
 }
+
+.badges {
+    font-size: 20px;
+}
